HuggingFace-Course/requirements.txt
markdownify
smolagents
requests
duckduckgo_search
pandas
pytz
wikipedia-api
newsapi-python
forex-python
sympy
python-dotenv
smolagents[gradio]